Update merge algorithm
We use `note.ver` to track changes in `note.text` (and tags and plugins
in the future).
Separately we need to update `note.nextReview` as it is possible that
older version of note was reviewed recently on one device but was edited
on the another one.
We use `note.lastReview` to merge `nextReview`.